Move object methods to the kernel
This allows TypeFactory to create the primordial 'type' and 'object'
type objects without access rights to the runtime module.
